+++ src/lib.rs
@@ -22,7 +22,7 @@ See [ChangeLog.md](https://github.com/kubo/rust-oracle/blob/master/ChangeLog.md)
 
 ## Build-time Requirements
 
-* Rust 1.19 or later
+* Rust 1.26 or later
 * C compiler. See `Compile-time Requirements` in [this document](https://github.com/alexcrichton/cc-rs#compile-time-requirements).
 
 ## Run-time Requirements